Start on moving handlers to own files
