+++ globals/vars.go
@@ -1,0 +1,3 @@
+package globals
+
+var OengusBotId = "559625844197163008"
